Allow tags to be marked as raw text
This fixes a bug related to markdown compilation where a code block
contains what the compiler things are expressions.

The fix here is to allow a `data-astro-raw` boolean attr to declare a
codeblock to be raw text. This attribute will be added in the markdown
plugin.

The old compiler did not have this problem since Markdown was compiled
inside of the compiler itself.